Add triangle to test app
Compile and run a shader to render a simple triangle. Simplest test
case. We briefly explored using old style glBegin(), glVertex3f() etc.
as a simpler case, however under the hood in mesa this turns into shader
compilation anyways. We might as well write our own shaders where we can
control the complexity

This comes with a patch for mesa to not crash when using our driver,
however there is no real implementation