Proof of concept shader compilation
Previously we updated our test app to run a full opengl pipeline. This
came with crashes due to unimplemented functions in mesa which we fixed.

Here we actually implement what needs to be done in those functions. At
a high level...

* Add shader compilation and execution to libgpu
* Add APIs on the gpu, and in the kernel to actually do this execution
* Link libgpu in mesa to do shader compilation

See patches for linux/qemu/mesa for more details on those

Libgpu details:

Shader compilation is implemented as a close mapping to the NIR shader
that is given to us by mesa. For a simple shader, not much needs to be
done other than loading/storing from constants/inputs.

GPU pipeline execution consists of
* Parsing inputs to the vertex shader according to rules defined by mesa
* Running vertex shader on those inputs
* Running the fragment shader (no inputs supported for now)
* Rasterizing triangles according to the outputs of the vertex shader,
  with the color from the fragment shader

In order to avoid having to design a bytecode with a huge lack of
information, we use json as our bytecode for now. We just take our
command enum array and json.serialize() him and json.parse() him when we
get to the GPU. This is the benefit of working entirely in software :)